Add animation for zoom level indicator
